% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{addNonNulls}
\alias{addNonNulls}
\title{Add all non-NULL elements of list.}
\usage{
addNonNulls(compList)
}
\arguments{
\item{compList}{List of numeric matrices or data.frames of the equal 
dimensions.}
}
\value{
Matrix or data.frame containing sum of all list elements where 
\code{is.null} is FALSE.
}
\description{
Add all non-NULL elements of list.
}
