% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{commaList2vector}
\alias{commaList2vector}
\title{Comma-separated string to numeric vector.}
\usage{
commaList2vector(commastring)
}
\arguments{
\item{commastring}{Input character vector containing numbers separated by 
commas.}
}
\value{
Numeric vector of values extracted from commastring.
}
\description{
Split input of comma-separated string into vector of numeric values.
}
