% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variancecomponentFunctions.R
\name{correlatedBgEffects}
\alias{correlatedBgEffects}
\title{Simulate correlated background effects.}
\usage{
correlatedBgEffects(N, P, pcorr)
}
\arguments{
\item{N}{number [integer] of samples to simulate}

\item{P}{number [integer] of phenotypes to simulate}

\item{pcorr}{initial strength of correlation [double] between neighbouring 
traits}
}
\value{
[N x P] matrix of correlated background effects
}
\description{
correlatedBgEffects computes a background effect that simulates structured 
correlation between the phenotypes.
}
\details{
correlatedBgEffects can be used to simulate phenotypes with a 
defined level of correlation between traits. The level of correlation depends
 on the distance
of the traits. Traits of distance d=1 (adjacent columns) will have 
correlation cor=\eqn{pcorr^1}{pcorr^1}, traits with d=2 have 
cor=\eqn{pcorr^2}{pcorr^2} up to traits with d=(P-1) 
cor=\eqn{pcorr^{(P-1)}}{pcorr^{(P-1)}}. The correlated background effect 
correlated is simulated based on this correlation structure C: 
\eqn{correlated ~ N_{NP}(0,C)}{correlated ~ N_{NP}(0,C)}.
}
\examples{
correlatedBg <- correlatedBgEffects(N=100, P=20, pcorr=0.4)
}
\seealso{
\code{\link[mvtnorm]{rmvnorm}} which is used to simulate the 
multivariate normal distribution
}
