% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createphenotypeFunctions.R
\name{createPheno}
\alias{createPheno}
\title{Combine different phenotype components.}
\usage{
createPheno(P, N, sampleID = "ID_", phenoID = "trait_",
  correlatedBg = NULL, genFixed = NULL, genBg = NULL, noiseFixed = NULL,
  noiseBg = NULL, genVar = NULL, h2s = NULL, h2bg = NULL,
  noiseVar = NULL, rho = NULL, delta = NULL, phi = NULL, gamma = 0.8,
  theta = 0.8, eta = 0.8, alpha = 0.8, pcorr = 0.6,
  modelNoise = "noNoise", modelGenetic = "noGenetic", verbose = TRUE)
}
\arguments{
\item{P}{number [integer] of phenotypes to simulate}

\item{N}{number [integer] of samples to simulate}

\item{sampleID}{prefix [string] for naming samples (followed by sample number
from 1 to N)}

\item{phenoID}{prefix [string] for naming traits (followed by trait number 
from 1 to P)}

\item{correlatedBg}{list of correlated background effects as obtained 
by \code{\link{correlatedBgEffects}}}

\item{genFixed}{list of independent and shared genetic fixed effects as 
obtained by \code{\link{geneticFixedEffects}}}

\item{genBg}{list of independent and shared genetic background effects as 
obtained by \code{\link{geneticBgEffects}}}

\item{noiseFixed}{list of independent and shared noise fixed effects as 
obtained by \code{\link{noiseFixedEffects}}}

\item{noiseBg}{list of independent and shared noise background effects as 
obtained by \code{\link{noiseBgEffects}}}

\item{genVar}{Proportion [double] of total genetic variance}

\item{h2s}{Proportion [double] of variance of fixed genetic effects}

\item{h2bg}{Proportion [double] of variance of background genetic effects; 
either h2s or h2b have to be specified and 
h2s + h2b = genVar}

\item{noiseVar}{Proportion [double] of total noise variance}

\item{rho}{Proportion [double] of variance of correlated noise effects}

\item{delta}{Proportion [double] of fixed noise variance}

\item{phi}{Proportion [double] of variance of background noise effects}

\item{gamma}{Proportion [double] of variance of shared fixed noise effects}

\item{theta}{Proportion [double] of variance of shared fixed genetic effects}

\item{eta}{Proportion [double] of variance of shared bg genetic effects}

\item{alpha}{Proportion [double] of variance of shared bg noise effect}

\item{pcorr}{Correlation [double] between phenotypes}

\item{modelNoise}{name [string] of noise model for the phenotype simulation; 
based on model independentation, phenotype components will be added to the  
final phenotype. Possible models: "noNoise", "noiseFixedOnly", "noiseBgOnly", 
"noiseCorrelatedOnly", "noiseFixedAndBg","noiseCorrelatedAndBg", 
"noiseFixedAndCorrelated", "noiseFixedAndBgAndCorrelated"}

\item{modelGenetic}{name [string] of genetic model for the phenotype 
simulation; based on model independentation, phenotype components will be 
added to the final phenotype. Possible models: "noGenetic","geneticBgOnly", 
"geneticFixedOnly","geneticFixedAndBg"}

\item{verbose}{[boolean]; if TRUE, progress info is printed to standard out}
}
\value{
named list of absolute levels of variance explained by each phenotype 
component (varComponents), 
the scaled phenotype components (phenoComponents) and and parameters used for 
phenotype set-up and labeling (setup)
}
\description{
createPheno takes precomputed phenotype components and rescales them 
according to the specified proportion of variance they should explain in the 
final phenotype.
}
\examples{
noiseBg <- noiseBgEffects(N=100, P=50)
phenotypeNoiseBgOnly <- createPheno(P=50, N=100, noiseBg=noiseBg, 
modelNoise="noiseBgOnly", modelGenetic="noGenetic")

genotypes <- simulateGenotypes(N=100, NrSNP=50)
causalSNPs <- getCausalSNPs(genotypes=genotypes, standardise=FALSE)
genFixed <- geneticFixedEffects(N=100, P=50, X_causal=causalSNPs)
phenotypeNoiseBgOnlyGenFixed <- createPheno(P=50, N=100, noiseBg=noiseBg,
genFixed=genFixed, modelNoise="noiseBgOnly", 
modelGenetic="geneticFixedOnly", genVar=0.05)
}
\seealso{
\code{\link{rescaleVariance}} for the function used to rescale 
phenotype components
}
