% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variancecomponentFunctions.R
\name{geneticBgEffects}
\alias{geneticBgEffects}
\title{Simulate genetic background effects.}
\usage{
geneticBgEffects(N, P, kinship)
}
\arguments{
\item{N}{number [integer] of samples to simulate}

\item{P}{number [integer] of phenotypes to simulate}

\item{kinship}{[N x N] matrix of kinship estimates [double]}
}
\value{
named list of common background genetic effects (common: [N x P] 
matrix) and specific background genetic effects (specific: [N x P] matrix)
}
\description{
geneticBgEffects simulates two random genetic effects (common and specific) 
based on the kinship estimates of the (simulated) samples.
}
\details{
For the simulation of the genetic background effects, three matrix 
components are used: i) the kinship matrix K [N x N] which is treated as the 
sample-design matrix
(the genetic profile of the samples), ii) an effect-size matrix B [N x P] 
with vec(B) drawn from a normal distribution and iii) the trait design matrix
 A [P x P]. For the
specific effect, A is a diagonal matrix with normally distributed values. A 
for the common effect is a matrix of rowrank one, with normally distributed 
entries in row 1 and
zeros elsewhere. The final effect E, the three matrices are multiplied as: 
E = KBA
}
\examples{
genotypes <- simulateGenotypes(N=100, NrSNP=400, verbose=FALSE)
kinship <- getKinship(genotypes$X_sd, norm=TRUE, verbose=FALSE)
geneticBg <- geneticBgEffects(N=100, P=10, kinship=kinship)
}
