% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandlineFunctions.R
\name{get_options}
\alias{get_options}
\title{Command line parser for PhenotypeSimulator}
\usage{
get_options()
}
\value{
args [list] of command line arguments
}
\description{
get_options runs without arguments. Upon call, it reads command-line
parameters via parse_args.
}
\examples{
# for help on usage of the command line interface run (on command line):
# (not run)
# Rscript --vanilla --default-packages=R.utils,PhenotypeSimulator,optparse
-e 'get_options()'  

}
