% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeFunctions.R
\name{getCausalSNPs}
\alias{getCausalSNPs}
\title{Draw random SNPs from genotypes.}
\usage{
getCausalSNPs(NrCausalSNPs = 20, genotypes = NULL, chr = NULL,
  chr_string = NULL, genoFilePrefix = NULL, genoFileSuffix = NULL,
  genoFileDelimiter = ",", sampleID = "ID_", standardise = FALSE,
  verbose = TRUE)
}
\arguments{
\item{NrCausalSNPs}{number [integer] of SNPs to chose at random}

\item{genotypes}{[NrSamples x totalNrSNPs] matrix of genotypes}

\item{chr}{numeric vector of chromosomes to chose NrCausalSNPs from; only 
used when external genotype
data is provided i.e. is.null(genoFilePrefix) == FALSE}

\item{chr_string}{[string] alternative to chr, a string with chromosomes 
separated by comma; 
most often used when run as a command line application}

\item{genoFilePrefix}{full path/to/chromosome-wise-genotype-file-ending-
before-"chrChromosomeNumber" (no '~' expansion!) [string]}

\item{genoFileSuffix}{[string] following chromosome number including 
.fileformat (e.g. ".csv")}

\item{genoFileDelimiter}{field separator [string] of genotype file}

\item{sampleID}{prefix [string] for naming samples (followed by sample number
from 1 to NrSamples)}

\item{standardise}{[boolean], if TRUE standardised genotypes will be returned}

\item{verbose}{[boolean]; if TRUE, progress info is printed to standard out}
}
\value{
[N x NrCausalSNPs] matrix of randomly drawn, standardised (depending 
on standardise option) SNPs
}
\description{
Draw random SNPs from either simulated genotypes or external genotype files.
}
\details{
In order to chose SNPs from external genotype files without reading 
them into memory, genotypes for each chromosome need to be accesible as 
[SNPs x samples] in a separate file, containing "chrChromosomenumber" (e.g 
chr22) in the file name (e.g. /path/to/dir/related_nopopstructure_chr22.csv).  
All genotype files need to be saved in the same directory. genoFilePrefix 
(/path/to/dir/related_nopopstructure_) and genoFileSuffix (.csv) specify the 
strings leading and following the "chrChromosomenumber". The first column in 
each file needs to be the SNP_ID and files cannot contain a header. 
getCausalSNPs generates a vector of chromosomes from which to sample the SNPs
. For each of the chromosomes, it counts the number of SNPs in the chromosome
 file and creates vectors of random numbers ranging from 1:NrSNPSinFile. Only 
 the lines corresponding to these numbers are then read into R.
}
\examples{
geno <- simulateGenotypes(N=500)
causalSNPsFromSimulatedGenoStandardised <- getCausalSNPs(genotypes=geno, 
standardise=TRUE)

genotypeFile <- system.file("extdata/genotypes/",
"related_nopopstructure_chr22.csv",
package = "PhenotypeSimulator")
genoFilePrefix <- gsub("chr.*", "", genotypeFile) 
genoFileSuffix <- ".csv" 
causalSNPsFromFile <- getCausalSNPs(chr=22, genoFilePrefix=genoFilePrefix, 
genoFileSuffix=genoFileSuffix)
}
\seealso{
\code{\link{standardiseGenotypes}}
}
