% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeFunctions.R
\name{getKinship}
\alias{getKinship}
\title{Get genetic kinship.}
\usage{
getKinship(X = NULL, kinshipfile = NULL, sampleID = "ID_", norm = FALSE,
  sep = ",", header = TRUE, verbose = TRUE)
}
\arguments{
\item{X}{[NrSamples x totalNrSNPs] matrix of standardised genotypes}

\item{kinshipfile}{path/to/kinshipfile [string] to be read; either X or 
kinshipfile must be provided}

\item{sampleID}{prefix [string] for naming samples (followed by sample number 
from 1 to NrSamples)}

\item{norm}{[boolean], if TRUE kinship matrix will be normalised by the mean 
of its diagonal elements
and 1e-4 added to the diagonal for numerical stability}

\item{sep}{field separator [string] of kinship file}

\item{header}{[boolean], if TRUE kinship file has header information}

\item{verbose}{[boolean]; if TRUE, progress info is printed to standard out}
}
\value{
[NrSamples x NrSamples] matrix of kinship estimate
}
\description{
Estimate kinship from standardised genotypes or read pre-computed kinship 
file. Standardised genotypes can be obtained via 
\code{\link{standardiseGenotypes}} or directly from the 
\code{\link{simulateGenotypes}} return object. 
Kinship matrices can optionally be normalised.
}
\examples{
geno <- simulateGenotypes(N=500)
K_fromGenotypesNormalised <- getKinship(geno$X_sd, norm=TRUE)

kinshipfile <- system.file("extdata/kinship", 
"kinship_unrelated_nopopstructure.csv",
package = "PhenotypeSimulator")
K_fromFile <- getKinship(kinshipfile=kinshipfile)
}
