% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variancecomponentFunctions.R
\name{noiseBgEffects}
\alias{noiseBgEffects}
\title{Simulate noise background effects.}
\usage{
noiseBgEffects(N, P, mean = 0, sd = 1)
}
\arguments{
\item{N}{number [integer] of samples to simulate}

\item{P}{number [integer] of phenotypes to simulate}

\item{mean}{mean [double] of the normal distribution}

\item{sd}{standard deviation [double] of the normal distribution}
}
\value{
named list of common background noise effects (common: [N x P] 
matrix) and specific background noise effects (specific: [N x P] matrix)
}
\description{
noiseBgEffects simulates two random noise effects (common and specific).
}
\details{
The common background effect common as simulated as vec(common) ~ 
N(mean,sd). The specific background effect sepcific is simulated as the 
matrix product of two normal
distributions A [N x 1] and B [P x 1]: specific At(B)
}
\examples{
noiseBG <- noiseBgEffects(N=100, P=20, mean=2)
}
