% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeSimulator.R
\docType{package}
\name{PhenotypeSimulator}
\alias{PhenotypeSimulator}
\title{PhenotypeSimulator: A package for simulating phenotypes from different
genetic and noise components}
\description{
Type, Name and Dependencies
}
\keyword{internal}
