% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_lines}
\alias{read_lines}
\title{Scan file for specific line numbers}
\usage{
read_lines(filename, lines, sep = "\\n")
}
\arguments{
\item{filename}{/path/to/chromosomefile [string]}

\item{lines}{vector of line numbers [integer] to be read}

\item{sep}{[string] end-of-line delimiter}
}
\description{
Scan file for specific line numbers
}
