% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createphenotypeFunctions.R
\name{rescaleVariance}
\alias{rescaleVariance}
\title{Scale phenotype component.}
\usage{
rescaleVariance(component, propvar)
}
\arguments{
\item{component}{numeric [N x P] phenotype matrix where N are the number of 
observations and P numer of phenotypes}

\item{propvar}{numeric specifying the proportion of variance that should be 
explained by this phenotype component}
}
\value{
component_scaled numeric [N x P] phenotype matrix if propvar != 0 or 
NULL
}
\description{
The function scales the specified phenotype component such that the average 
column variance is equal
to the user-specified proportion of variance.
}
\examples{
x <- matrix(rnorm(100), nc=10)
x_scaled <- rescaleVariance(x, propvar=0.4)
}
