% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createphenotypeFunctions.R
\name{setModel}
\alias{setModel}
\title{Set simulation model.}
\usage{
setModel(genVar = NULL, h2s = NULL, theta = 0.8, h2bg = NULL,
  eta = 0.8, noiseVar = NULL, delta = NULL, gamma = 0.8, rho = NULL,
  phi = NULL, alpha = 0.8, pcorr = 0.6, pIndependentConfounders = 0.4,
  pTraitIndependentConfounders = 0.2, pIndependentGenetic = 0.4,
  pTraitIndependentGenetic = 0.2, v = TRUE)
}
\arguments{
\item{genVar}{Total genetic variance [double]}

\item{h2s}{Proportion [double] of variance of fixed genetic effects}

\item{theta}{Proportion [double] of variance of shared fixed genetic effects}

\item{h2bg}{Proportion [double] of variance of random genetic effects}

\item{eta}{Proportion [double] of variance of shared bg genetic effects}

\item{noiseVar}{Total genetic variance [double]}

\item{delta}{Proportion [double] of fixed noise variance}

\item{gamma}{Proportion [double] of variance of shared fixed noise effects}

\item{rho}{Proportion [double] of variance of correlated noise effects}

\item{phi}{Proportion [double] of variance of background noise effects}

\item{alpha}{Proportion [double] of Variance of shared bg noise effect}

\item{pcorr}{Correlation [double] between phenotypes}

\item{pIndependentConfounders}{Proportion [double] of noise effects 
(confounders) to have a trait-independent effect}

\item{pTraitIndependentConfounders}{Proportion [double] of traits influenced
by independent fixed noise effects}

\item{pIndependentGenetic}{Proportion [double] of genetic effects (SNPs) to 
have a trait-independent fixed effect}

\item{pTraitIndependentGenetic}{Proportion [double] of traits influenced by 
independent fixed genetic effects}

\item{v}{[boolean]; if TRUE, progress info is printed to standard out}
}
\value{
named list containing the genetic model (modelGenetic) and the noise 
model (modelNoise). Options are:
modelNoise: "nNoise", "noiseFixedOnly", "noiseBgOnly", "noiseCorrelatedOnly",
 "noiseFixedAndBg","noiseCorrelatedAndBg", "noiseFixedAndCorrelated",
 "noiseFixedAndBgAndCorrelated"
modelGenetic: "noGenetic","geneticBgOnly", "geneticFixedOnly",
"geneticFixedAndBg"
}
\description{
Based on parameters provided, this function sets the name for the phenotype 
simulation. The model name is needed downstream for phenotype component 
simulations, but can also be set manually.
}
\examples{
#genetic fixed effects only
model <- setModel(genVar=1, h2s=1)

#genetic fixed and bg effects
model <- setModel(genVar=1, h2s=0.01)

#genetic and noise fixed effects only
model <- setModel(genVar=0.4, h2s=1, delta=1)
}
