% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeFunctions.R
\name{simulateGenotypes}
\alias{simulateGenotypes}
\title{Simulate bi-allelic genotypes.}
\usage{
simulateGenotypes(N, NrSNP = 5000, frequencies = c(0.1, 0.2, 0.4),
  sampleID = "ID_", verbose = TRUE)
}
\arguments{
\item{N}{number of samples for which to simulate bi-allelelic genotypes}

\item{NrSNP}{number of SNPs to simulate}

\item{frequencies}{vector of allele frequencies [double] from which to sample}

\item{sampleID}{prefix [string] for naming samples (followed by sample number 
from 1 to N)}

\item{verbose}{boolean; if TRUE, progress info is printed to standard out}
}
\value{
list of a [NrSNP x N] matrix of simulated genotypes, SNP frequencies 
and vector of sample IDs.
}
\description{
Simulate bi-allelic genotypes.
}
\examples{
N10NrSNP10 <- simulateGenotypes(N=10, NrSNP=10)
N10NrSNP10 <- simulateGenotypes(N=10, NrSNP=10,
frequency="0.2,0.3,0.4")
}
\seealso{
\code{\link{standardiseGenotypes}}
}
