% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeFunctions.R
\name{standardiseGenotypes}
\alias{standardiseGenotypes}
\title{Standardise genotypes}
\usage{
standardiseGenotypes(geno)
}
\arguments{
\item{geno}{[N x NrSNP] matrix/dataframe of genotypes}
}
\value{
[N x NrSNP] matrix of standardised genotypes
}
\description{
Genotypes are standardised as described in Yang et al:
snp_standardised = (snp - 2 * minor_allele_freq)/
sqrt(2 * minor_allele_freq * major_allele_freq)
}
\examples{
geno <- cbind(rbinom(100, 2, 0.3), rbinom(100, 2, 0.4))
geno_sd <- standardiseGenotypes(geno)
}
\references{
Yang, J., Lee, S.H., Goddard, M.E., Visscher, P.M. (2011) GCTA: 
a tool for genome-wide complex trait analysis, AJHG: 88
}
\seealso{
\code{\link{getAlleleFrequencies}}
}
