% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variancecomponentFunctions.R
\name{geneticFixedEffects}
\alias{geneticFixedEffects}
\title{Simulate genetic fixed effects.}
\usage{
geneticFixedEffects(X_causal, N, P, pSpecificGenetic = 0.4,
  pTraitSpecificGenetic = 0.2)
}
\arguments{
\item{X_causal}{[N x NrCausalSNPs] matrix of standardised (depending on 
standardise option) SNPs}

\item{N}{number [integer] of samples to simulate}

\item{P}{number [integer] of phenotypes to simulate}

\item{pSpecificGenetic}{Proportion [double] of genetic effects (SNPs) to have
a trait-specific fixed effect}

\item{pTraitSpecificGenetic}{Proportion [double] of traits influenced by 
specific fixed genetic effects}
}
\value{
named list of common fixed genetic effects (common: [N x P] matrix), 
specific fixed genetic effects (specific: [N x P] matrix), 
the causal SNPs named by having a common or specific effect 
(cov: [NrCausalSNPs x N] matrix) and the simulated effect sizes of the causal 
SNPs (cov_effect: [P x NrCausalSNPs] dataframe)
}
\description{
geneticFixedEffects takes a matrix of SNPs which should be added as fixed 
effect to the phenotype. SNPs can have effects across all traits (common) 
or to a number of traits only (specific); the proportion of specific SNPs 
from the total input SNPs can be chosen via pSpecificGenetic. The number of 
traits 
that are associated with specific genetic effects can be chosen via 
pTraitSpecificGenetic.
}
\examples{
genotypes <- simulateGenotypes(N=100, NrSNP=20, verbose=FALSE)
causalSNPs <- getCausalSNPs(genotypes=genotypes)
geneticFixed <- geneticFixedEffects(X_causal=causalSNPs, P=10, N=100)
}
