% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variancecomponentFunctions.R
\name{geneticFixedEffects}
\alias{geneticFixedEffects}
\title{Simulate genetic fixed effects.}
\usage{
geneticFixedEffects(X_causal, P, N = NULL, pIndependentGenetic = 0.4,
  pTraitIndependentGenetic = 0.2, verbose = TRUE)
}
\arguments{
\item{X_causal}{[N x NrCausalSNPs] matrix of standardised (depending on 
standardise option) SNPs}

\item{P}{number [integer] of phenotypes to simulate}

\item{N}{number [integer] of samples to simulate; has to be less than or 
equal to the number of samples in X_causal (rows); if less than  number of 
samples in X_causal, N random rows of X_causal will be drawn; if not 
specified, N assumed to be equal to samples in X_causal}

\item{pIndependentGenetic}{Proportion [double] of genetic effects (SNPs) to
have an independent fixed effect}

\item{pTraitIndependentGenetic}{Proportion [double] of traits influenced by 
independent fixed genetic effects}

\item{verbose}{[boolean]; if TRUE, progress info is printed to standard out}
}
\value{
named list of shared fixed genetic effects (shared: [N x P] matrix), 
independent fixed genetic effects (independent: [N x P] matrix), 
the causal SNPs named by having a shared or independent effect 
(cov: [NrCausalSNPs x N] matrix) and the simulated effect sizes of the causal 
SNPs (cov_effect: [P x NrCausalSNPs] dataframe)
}
\description{
geneticFixedEffects takes a matrix of SNPs which should be added as fixed 
effect to the phenotype. SNPs can have effects across all traits (shared) 
or to a number of traits only (independent); the proportion of independent 
SNPs from the total input SNPs can be chosen via pIndependentGenetic. The 
number of traits that are associated with independent genetic effects can be 
chosen via pTraitIndependentGenetic.
}
\examples{
genotypes <- simulateGenotypes(N=100, NrSNP=20, verbose=FALSE)
causalSNPs <- getCausalSNPs(genotypes=genotypes)
geneticFixed <- geneticFixedEffects(X_causal=causalSNPs, P=10, N=100)
}
