% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeFunctions.R
\name{getCausalSNPs}
\alias{getCausalSNPs}
\title{Draw random SNPs from genotypes.}
\usage{
getCausalSNPs(NrCausalSNPs = 20, genotypes = NULL, chr = NULL,
  NrChrCausal = NULL, genoFilePrefix = NULL, genoFileSuffix = NULL,
  delimiter = ",", sampleID = "ID_", verbose = TRUE)
}
\arguments{
\item{NrCausalSNPs}{number [integer] of SNPs to chose at random}

\item{genotypes}{[NrSamples x totalNrSNPs] matrix of genotypes}

\item{chr}{numeric vector of chromosomes to chose NrCausalSNPs from; only 
used when external genotype data is provided i.e. is.null(genoFilePrefix) 
== FALSE}

\item{NrChrCausal}{Number [integer] of causal chromosomes to  chose 
NrCausalSNPs from (as opposed to the actual chromosomes to chose from via chr
);  only used when external genotype data is provided i.e. 
is.null(genoFilePrefix) == FALSE.}

\item{genoFilePrefix}{full path/to/chromosome-wise-genotype-file-ending-
before-"chrChromosomeNumber" (no '~' expansion!) [string]}

\item{genoFileSuffix}{[string] following chromosome number including 
.fileformat (e.g. ".csv"); has to be a text format i.e. comma/tab/space
separated}

\item{delimiter}{field separator [string] of genotypefile or genoFile}

\item{sampleID}{prefix [string] for naming samples (followed by sample number
from 1 to NrSamples)}

\item{verbose}{[boolean]; if TRUE, progress info is printed to standard out}
}
\value{
[N x NrCausalSNPs] matrix of randomly drawn, standardised (depending 
on standardise option) SNPs
}
\description{
Draw random SNPs from either simulated genotypes or external genotype files.
When drawing from external genotype files, two options are possible: 
i) reading the entire file into memory or ii) only reading lines of randomly
chosen SNPs, recommended for large genotype files. See details for more 
information.
}
\details{
In order to chose SNPs from external genotype files without reading 
them into memory, genotypes for each chromosome need to be accesible as 
[SNPs x samples] in a separate file, containing "chrChromosomenumber" (e.g 
chr22) in the file name (e.g. /path/to/dir/related_nopopstructure_chr22.csv).  
All genotype files need to be saved in the same directory. genoFilePrefix 
(/path/to/dir/related_nopopstructure_) and genoFileSuffix (.csv) specify the 
strings leading and following the "chrChromosomenumber". The first column in 
each file needs to be the SNP_ID and files cannot contain a header. 
getCausalSNPs generates a vector of chromosomes from which to sample the SNPs
. For each of the chromosomes, it counts the number of SNPs in the chromosome
 file and creates vectors of random numbers ranging from 1:NrSNPSinFile. Only 
 the lines corresponding to these numbers are then read into R. The example 
 data provided for chromosome 22 contains genotypes (50 samples) of the first 
 500 SNPs on chromosome 22 with a minor allele frequency of greater than 2% 
 from the European populations of the the 1000 Genomes project.
}
\examples{
# get causal SNPs from genotypes simulated within PhenotypeSimulator
geno <- simulateGenotypes(N=10, NrSNP=10)
causalSNPsFromSimulatedGenoStandardised <- getCausalSNPs(NrCausalSNPs=10,
genotypes=geno$genotypes)

# Get causal SNPs by sampling lines from large SNP files
genotypeFile <- system.file("extdata/genotypes/",
"genotypes_chr22.csv",
package = "PhenotypeSimulator")
genoFilePrefix <- gsub("chr.*", "", genotypeFile) 
genoFileSuffix <- ".csv" 
causalSNPsFromLines <- getCausalSNPs(NrCausalSNPs=10, chr=22, 
genoFilePrefix=genoFilePrefix, 
genoFileSuffix=genoFileSuffix)
}
\seealso{
\code{\link{standardiseGenotypes}}
}
