% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variancecomponentFunctions.R
\name{noiseBgEffects}
\alias{noiseBgEffects}
\title{Simulate noise background effects.}
\usage{
noiseBgEffects(N, P, mean = 0, sd = 1)
}
\arguments{
\item{N}{number [integer] of samples to simulate}

\item{P}{number [integer] of phenotypes to simulate}

\item{mean}{mean [double] of the normal distribution}

\item{sd}{standard deviation [double] of the normal distribution}
}
\value{
named list of shared background noise effects (shared: [N x P] 
matrix) and independent background noise effects (independent: [N x P] 
matrix)
}
\description{
noiseBgEffects simulates two random noise effects (shared and independent).
}
\details{
The independent background effect is simulated as vec(shared) ~ 
N(mean,sd). The shared background effect is simulated as the 
matrix product of two normal
distributions A [N x 1] and B [P x 1]: AB^t
}
\examples{
noiseBG <- noiseBgEffects(N=100, P=20, mean=2)
}
