% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createphenotypeFunctions.R
\name{runSimulation}
\alias{runSimulation}
\title{Run phenotype simulation.}
\usage{
runSimulation(N = 1000, P = 10, tNrSNP = 5000, cNrSNP = 20,
  NrConfounders = 10, seed = 219453, chr_string = NULL, chr = NULL,
  NrChrCausal = NULL, genVar = NULL, h2s = NULL, theta = 0.8,
  h2bg = NULL, eta = 0.8, noiseVar = NULL, rho = NULL, delta = NULL,
  gamma = 0.8, phi = NULL, alpha = 0.8, sampleID = "ID_",
  phenoID = "Trait_", genoFilePrefix = NULL, genoFileSuffix = NULL,
  genoFileDelimiter = ",", kinshipfile = NULL, kinshipHeader = TRUE,
  kinshipDelimiter = ",", normalise = TRUE, standardise = TRUE,
  NrFixedEffects = 1, distConfounders = "norm", mConfounders = 0,
  sdConfounders = 1, catConfounders = NULL, probConfounders = NULL,
  distBetaConfounders = "norm", mBetaConfounders = 0,
  sdBetaConfounders = 1, distBetaGenetic = "norm", mBetaGenetic = 0,
  sdBetaGenetic = 1, pIndependentConfounders = 0.4,
  pTraitIndependentConfounders = 0.2, pcorr = 0.8, meanNoiseBg = 0,
  sdNoiseBg = 1, SNPfrequencies = c(0.1, 0.2, 0.4),
  SNPfrequencyString = NULL, pIndependentGenetic = 0.4,
  pTraitIndependentGenetic = 0.2, NrConfoundersString = NULL,
  pIndependentConfoundersString = NULL,
  pTraitIndependentConfoundersString = NULL, distConfoundersString = NULL,
  mConfoundersString = NULL, sdConfoundersString = NULL,
  catConfoundersString = NULL, probConfoundersString = NULL,
  distBetaConfoundersString = NULL, mBetaConfoundersString = NULL,
  sdBetaConfoundersString = NULL, distBetaGeneticString = NULL,
  mBetaGeneticString = NULL, sdBetaGeneticString = NULL, verbose = TRUE)
}
\arguments{
\item{N}{number [integer] of samples to simulate}

\item{P}{number [integer] of phenotypes to simulate}

\item{tNrSNP}{total number [integer] of SNPs to simulate; these SNPs are used
for kinship estimation}

\item{cNrSNP}{number [integer] of causal SNPs; used as genetic fixed effects}

\item{NrConfounders}{number [integer] of confounders; used as noise fixed 
effects}

\item{seed}{seed [integer] to initiate random number generation}

\item{chr_string}{alternative to chr, a [string] with chromosomes separated 
by comma; most often used when run as a command line application.}

\item{chr}{numeric vector of chromosomes to chose NrCausalSNPs from; only 
used when external genotype data is provided i.e. is.null(genoFilePrefix) == 
FALSE}

\item{NrChrCausal}{Number [integer] of causal chromosomes to  chose 
NrCausalSNPs from (as opposed to the actual chromosomes to chose from via chr
'chr_string);  only used when external genotype data is provided i.e. 
is.null(genoFilePrefix) == FALSE.}

\item{genVar}{Proportion [double] of total genetic variance}

\item{h2s}{Proportion [double] of gentic variance of fixed effects}

\item{theta}{Proportion [double] of variance of shared fixed genetic effects}

\item{h2bg}{Proportion [double] of genetic variance of background effects; 
either h2s or h2bg have to be specified and 
h2s + h2bg = 1}

\item{eta}{Proportion [double] of variance of shared bg genetic effects}

\item{noiseVar}{Proportion [double] of total noise variance}

\item{rho}{Proportion [double] of noise variance of correlated effects; sum 
of rho, delta and phi cannot be greater than 1}

\item{delta}{Proportion [double] of noise variance of fixed effects; sum of 
rho, delta and phi cannot be greater than 1}

\item{gamma}{Proportion [double] of variance of shared fixed noise effects}

\item{phi}{Proportion [double] of noise variance of background effects; sum 
of rho, delta and phi cannot be greater than 1}

\item{alpha}{Variance [double] of shared bg noise effect}

\item{sampleID}{prefix [string] for naming samples (followed by sample number
from 1 to N)}

\item{phenoID}{prefix [string] for naming traits (followed by trait number 
from 1 to P)}

\item{genoFilePrefix}{path/to/chromosome-wise-genotype-file-ending-before-
"chrChromosomeNumber" [string]}

\item{genoFileSuffix}{[string] following chromosome number including 
.fileformat (e.g. ".csv")}

\item{genoFileDelimiter}{field separator [string] of genotype file}

\item{kinshipfile}{path/to/kinshipfile [string] to be read; either X or 
kinshipfile must be provided}

\item{kinshipHeader}{[boolean], if TRUE kinship file has header information}

\item{kinshipDelimiter}{field separator [string] of kinship file}

\item{normalise}{[boolean], if TRUE kinship matrix will be normalised by the 
mean of its diagonal elements and 1e-4 added to the diagonal for numerical 
stability}

\item{standardise}{[boolean], if TRUE standardised genotypes will be returned}

\item{NrFixedEffects}{number [integer] of different fixed effects to simulate
; allows to simulate fixed effects from different distributions or with 
 different parameters}

\item{distConfounders}{name [string] of distribution to use to simulate 
confounders; one of "unif", "norm", "bin", "cat_norm", "cat_unif"}

\item{mConfounders}{mean/midpoint [double] of normal/uniform distribution for
confounders}

\item{sdConfounders}{standard deviation/extension from midpoint [double] of 
normal/uniform distribution for confounders}

\item{catConfounders}{confounder categories [factor]; required if 
distConfounders "cat_norm" or "cat_unif"}

\item{probConfounders}{probability [double] of binomial confounders (0/1); 
required if distConfounders "bin"}

\item{distBetaConfounders}{name [string] of distribution to use to simulate 
effect sizes of confounders; one of "unif" or "norm"}

\item{mBetaConfounders}{mean/midpoint [double] of normal/uniform distribution 
for effect sizes of confounders}

\item{sdBetaConfounders}{standard deviation/extension from midpoint [double] 
of normal/uniform distribution for effect sizes of confounders}

\item{distBetaGenetic}{name [string] of distribution to use to simulate 
effect sizes of SNPs; one of "unif" or "norm"}

\item{mBetaGenetic}{mean/midpoint [double] of normal/uniform distribution 
for effect sizes of SNPs}

\item{sdBetaGenetic}{standard deviation/extension from midpoint [double] 
of normal/uniform distribution for effect sizes of SNPs}

\item{pIndependentConfounders}{Proportion [double] of noise effects 
(confounders) to have a trait-independent effect}

\item{pTraitIndependentConfounders}{Proportion [double] of traits influenced 
by independent fixed noise effects}

\item{pcorr}{Correlation [double] between phenotypes}

\item{meanNoiseBg}{mean [double] of the normal distribution for noise bg 
effects}

\item{sdNoiseBg}{standard deviation [double] of the normal distribution for 
noise bg effects}

\item{SNPfrequencies}{vector of allele frequencies [double] from which to 
sample}

\item{SNPfrequencyString}{alternative to a frequencies vector, a [string] 
with frequencies separated by comma}

\item{pIndependentGenetic}{Proportion [double] of genetic effects (SNPs) to 
have a trait-independent fixed effect}

\item{pTraitIndependentGenetic}{Proportion [double] of traits influenced by 
independent fixed genetic effects}

\item{NrConfoundersString}{alternative to NrConfounder, a comma-separated
[string] with number(s) [integer] of confounders to simulate; typically used 
when run as command line application}

\item{pIndependentConfoundersString}{alternative to pIndependentConfounders, 
a comma-separated [string] with proportion(s) [double] of noise effects 
(confounders) to have a trait-independent effect; typically used when run as 
command line application}

\item{pTraitIndependentConfoundersString}{alternative to 
pTraitIndependentConfounders,  a comma-separated [string] with proportion(s) 
[double] of traits influenced by independent fixed noise effects; typically 
used when run as command line application}

\item{distConfoundersString}{alternative to distConfounders, a comma-
separated [string] with name(s) [string] of distributions to use to 
simulate confounders; one of "unif", "norm", "bin", "cat_norm", "cat_unif";
typically used when run as command line application}

\item{mConfoundersString}{alternative to mConfounders, a comma-
separated [string] with of mean/midpoint(s) [double] of normal/uniform 
distribution for confounders; typically used when run as command line 
application}

\item{sdConfoundersString}{alternative to sdConfounders, a comma-
separated [string] with standard deviation(s)/distance from 
midpoint(s) [double] of normal/uniform distribution for confounders; 
typically used when run as command line application}

\item{catConfoundersString}{alternative to catConfounders, a comma-
separated [string] with the number of confounder categories [integer]; 
required if distConfounders "cat_norm" or "cat_unif"; 
typically used when run as command line application}

\item{probConfoundersString}{alternative to probConfounders, a comma-
separated [string] with probability(s) [double] of binomial 
confounders (0/1); required if distConfounders "bin"; typically used
 when run as command line application}

\item{distBetaConfoundersString}{alternative to distBeta, a comma-separated 
[string] with name(s) [string] of distribution to use to simulate effect 
sizes of confounders; one of "unif" or "norm"; typically used when run as 
command line application}

\item{mBetaConfoundersString}{alternative to mBeta, a comma- separated 
[string] with means/midpoints [double] of normal/uniform distribution for 
effect sizes of confounders; typically used when run as command line 
application}

\item{sdBetaConfoundersString}{alternative to sdBeta, a comma- separated 
[string] with standard deviation/distance from midpoint [double] of 
normal/uniform 
distribution for effect sizes of confounders; typically used when run as 
command line application}

\item{distBetaGeneticString}{name [string] of distribution to use to simulate 
effect sizes of SNPs; one of "unif" or "norm"; typically used when run as 
command line application}

\item{mBetaGeneticString}{mean/midpoint [double] of normal/uniform 
distribution for effect sizes of SNPs; typically used when run as command 
line application}

\item{sdBetaGeneticString}{standard deviation/extension from midpoint 
[double] of normal/uniform distribution for effect sizes of SNPs; typically 
used when run as command line application}

\item{verbose}{[boolean]; if TRUE, progress info is printed to standard out}
}
\value{
named list of absolute levels of variance explained by each phenotype 
component (varComponents), 
the scaled phenotype components (phenoComponents) and and parameters used for 
phenotype set-up and labeling (setup)
}
\description{
runSimulation wraps around the phenotype component functions (genFixedEffects
, genBgEffects, noiseBgEffects, noiseFixedEffects and correlatedBgEffects)
and combines the simulated phenotype components via create phenotype.
}
\examples{
# simulate phenotype of 100 samples, 10 traits from genetic and noise 
# background effects, with variance explained of 0.2 and 0.8 respectively
genVar = 0.2
simulatedPhenotype <- runSimulation(N=100, P=5, cNrSNP=10,
genVar=genVar, h2s=1, phi=1)
}
\seealso{
\code{\link{createPheno}}, which this function wraps
}
