% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createphenotypeFunctions.R
\name{savePheno}
\alias{savePheno}
\title{Save final phenotype and phenotype components.}
\usage{
savePheno(simulatedData, directoryGeno, directoryPheno, kinship = NULL,
  genotypes = NULL, causalSNPs = NULL, sample_subset_vec = NULL,
  pheno_subset_vec = NULL, sample_subset_string = NULL,
  pheno_subset_string = NULL, outstring = NULL, saveAsTable = TRUE,
  saveAsRDS = FALSE, verbose = TRUE)
}
\arguments{
\item{simulatedData}{named list of i) dataframe of proportion of variance 
explained for each component (varComponents), 
ii) a named list with the simulated phenotype components (phenoComponents) 
and iii) a named list of parameters describing the model setup (setup);
obtained from either \link{runSimulation} or \link{createPheno}}

\item{directoryGeno}{name of parent directory [string] where genotypes from 
simulations should be saved [needs user writing permission]}

\item{directoryPheno}{name of parent directory [string] where final phenotype 
and phenotype components from simulations should be saved [needs user writing 
permission]}

\item{kinship}{[N x N] matrix of kinship estimate; optional: if provided, 
kinship estimate will be saved to file}

\item{genotypes}{[N x totalNrSNPs] matrix of all simulated SNPs; optional: if
provided, simulated SNPs will be saved to file}

\item{causalSNPs}{[N x NrCausalSNPs] matrix of causal SNPs; optional: if 
provided, causalSNPs will be saved to file}

\item{sample_subset_vec}{optional vector of sample subset sizes [integer];
if provided, draws subsets of samples out of the total simulated dataset and 
saves them separately}

\item{pheno_subset_vec}{optional vector of phenotype subset sizes [integer] 
if provided, draws subsets of traits out of the total simulated dataset 
and saves them separately}

\item{sample_subset_string}{optional [string] of comma-separated numbers e.g.
"50,100,500"; alternative to \code{sample_subset_vec}, typically used when
run as command line application}

\item{pheno_subset_string}{optional [string] of comma-separated numbers e.g. 
"10,50,80"; alternative to \code{pheno_subset_vec}, typically used when
 run as command line application}

\item{outstring}{optional name [string] of subdirectory (in relation to 
directoryPheno/directoryGeno) to save set-up
independent simulation results}

\item{saveAsTable}{[boolean] if TRUE, data will be saved as .csv}

\item{saveAsRDS}{[boolean] if TRUE, data will be saved as .RDS; at least one 
of 'saveAsTable' or 'saveAsRDS' has to be TRUE, both can be TRUE}

\item{verbose}{[boolean]; if TRUE, progress info is printed to standard out}
}
\value{
no return value, sole purpose of function is writing output files
}
\description{
savePheno saves model setup parameters and simulated genotypes to the 
specified directories. Requires a simulatedData list which is the output of 
either \link{runSimulation} or \link{createPheno}
}
\examples{
simulatedPhenotype <- runSimulation(N=100, P=10, genVar=0.4, h2bg=1, 
phi=1, verbose=FALSE)
#not run
#savePheno(simulatedPhenotype, directoryGeno="/path/to/dir/",  
#directoryPheno="/path/to/dir/", outstring="Date_simulation")
}
