% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createphenotypeFunctions.R
\name{savePheno}
\alias{savePheno}
\title{Save final phenotype and phenotype components.}
\usage{
savePheno(simulatedData, directoryGeno, directoryPheno,
  sample_subset_vec = NULL, pheno_subset_vec = NULL,
  sample_subset_string = NULL, pheno_subset_string = NULL,
  outstring = NULL, saveAsTable = TRUE, saveAsRDS = FALSE,
  saveAsPlink = FALSE, verbose = TRUE)
}
\arguments{
\item{simulatedData}{named list of i) dataframe of proportion of variance 
explained for each component (varComponents), 
ii) a named list with the simulated phenotype components (phenoComponents) 
and iii) a named list of parameters describing the model setup (setup);
obtained from either \link{runSimulation} or \link{createPheno}}

\item{directoryGeno}{absolute path (no tilde expansion) to parent directory 
[string] where genotypes from simulations should be saved [needs user writing 
permission]}

\item{directoryPheno}{absolute path (no tilde expansion) to parent directory 
[string] where final phenotype and phenotype components from simulations 
should be saved [needs user writing permission]}

\item{sample_subset_vec}{optional vector of sample subset sizes [integer];
if provided, draws subsets of samples out of the total simulated dataset and 
saves them separately}

\item{pheno_subset_vec}{optional vector of phenotype subset sizes [integer] 
if provided, draws subsets of traits out of the total simulated dataset 
and saves them separately}

\item{sample_subset_string}{optional [string] of comma-separated numbers e.g.
"50,100,500"; alternative to \code{sample_subset_vec}, typically used when
run as command line application}

\item{pheno_subset_string}{optional [string] of comma-separated numbers e.g. 
"10,50,80"; alternative to \code{pheno_subset_vec}, typically used when
 run as command line application}

\item{outstring}{optional name [string] of subdirectory (in relation to 
directoryPheno/directoryGeno) to save set-up
independent simulation results}

\item{saveAsTable}{[boolean] if TRUE, data will be saved as .csv}

\item{saveAsRDS}{[boolean] if TRUE, data will be saved as .RDS; at least one 
of 'saveAsTable' or 'saveAsRDS' has to be TRUE, both can be TRUE}

\item{saveAsPlink}{[boolean] if TRUE, simulated genotype data will be 
additionally be saved in binary PLINK format: .bed, .bim and .fam}

\item{verbose}{[boolean]; if TRUE, progress info is printed to standard out}
}
\value{
list of paths [strings] to final output phenotype (directoryPheno) 
and genotype (directoryGeno) directories. If outstring or subset settings not
NULL, these directories will be subdirectories of the input phenotype and 
genotype directories.
}
\description{
savePheno saves model setup parameters and simulated genotypes to the 
specified directories. Requires a simulatedData list which is the output of 
either \link{runSimulation} or \link{createPheno}
}
\examples{
simulatedPhenotype <- runSimulation(N=100, P=5, cNrSNP=10,
genVar=0.2, h2s=1, phi=1)
#not run
#outputdir <- savePheno(simulatedPhenotype, directoryGeno="/path/to/dir/",  
#directoryPheno="/path/to/dir/", outstring="Date_simulation", 
#saveAsPlink=TRUE)
}
