% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{writeStandardOutput}
\alias{writeStandardOutput}
\title{Write simulated data into formats used by standard GWAS software}
\usage{
writeStandardOutput(data, type, directory, id_samples, id_snps, id_traits,
  outstring = NULL, standardInput = NULL, format = c("plink", "snptest",
  "gemma", "bimbam", "delim"), intercept_gemma = FALSE,
  pheno_snptest = NULL, verbose = TRUE, delimiter = ",")
}
\arguments{
\item{data}{data.frame/matrix}

\item{type}{[string]}

\item{directory}{absolute path (no tilde expansion) to parent directory 
[string] where the data should be saved [needs user writing permission]}

\item{id_samples}{[string]}

\item{id_snps}{[string]}

\item{id_traits}{[string]}

\item{outstring}{optional name [string] of subdirectory (in relation to 
directoryPheno/directoryGeno) to save set-up
independent simulation results}

\item{standardInput}{[]}

\item{format}{name [string] of genotype file format}

\item{intercept_gemma}{[boolean] When modeling an intercept term in gemma, a 
column of 1's have to be appended to the covariate files. Set intercept_gemma
to TRUE to include a column of 1's in the output.}

\item{pheno_snptest}{The snptest .samples file contains sample IDs, 
covariates and phenotypes. Thus, when writing the covariates file, the 
formated sample ID and phenotype data have to be provided 
(output from type = "pheno", format ="snptest")}

\item{verbose}{[boolean]; if TRUE, progress info is printed to standard out}

\item{delimiter}{field separator [string] of genotype file}
}
\value{
List of [NrSamples X NrSNPs] genotypes, their [NrSNPs] ID (id_snps), 
their [NrSamples] IDs (id_samples) and format specific additional files (
might be used for output writing).
}
\description{
writeStandardOutput can write genotypes and phenotypes as well as possible
covariates and kinship matrices into a number of formats for standard GWAS: 
plink, snptest, bimbam, gemma. Alternatively, simple text files (
with specified delimiter) can be written. For more information on the 
different file formats see \emph{Externa formats}.
}
\section{External formats}{

\itemize{
\item PLINK: consists of three files, .bed, .bim and .fam. 
From \url{https://www.cog-genomics.org/plink/1.9/formats}:  The .bed 
files contain the  primary representation of genotype calls at biallelic 
variants in a binary format. The .bim is a text file with no header 
line, and one line  per variant with the following six fields: i) Chromosome 
code (either an integer, or 'X'/'Y'/'XY'/'MT'; '0' indicates unknown) or 
name, ii) Variant identifier, iii) Position in morgans or centimorgans (safe 
to use dummy value of '0'), iv) Base-pair coordinate (normally 1-based, but 
0 ok; limited to 231-2), v) Allele 1 (corresponding to clear bits in .bed; 
usually minor), vi) Allele 2 (corresponding to set bits in .bed; usually 
major). The .fam file is a text file with no header line, and one line per 
sample with the following six fields: i) Family ID ('FID'), ii), Within-
family ID ('IID'; cannot be '0'), iii) Within-family ID of father ('0' if 
father isn't in dataset, iv) within-family ID of mother ('0' if mother isn't 
in dataset), v) sex code ('1' = male, '2' = female, '0' = unknown), vi) 
Phenotype value ('1' = control, '2' = case, '-9'/'0'/non-numeric = missing 
data if case/control)
\item snptest: consists of two files, the genotype file ending in .gen 
and the sample file ending in .sample. From
 \url{http://www.stats.ox.ac.uk/~marchini/software/gwas/file_format.html}:
The genotype file stores data on a one-line-per-SNP format. The first 5 
entries of each line should be the SNP ID, RS ID of the SNP, base-pair 
position of the SNP, the allele coded A and the allele coded B. The SNP ID 
can be used to denote the chromosome number of each SNP. The next three 
numbers on the line should be the probabilities of the three genotypes AA, 
AB and BB at the SNP for the first individual in the cohort. The next three 
numbers should be the genotype probabilities for the second individual in the 
cohort. The next three numbers are for the third individual and so on. The 
order of individuals in the genotype file should match the order of the 
individuals in the sample file. The sample file has three parts (a) a header 
line detailing the names of the columns in the file, (b) a line detailing the 
types of variables stored in each column, and (c) a line for each individual 
detailing the information for that individual. For more information on the 
sampple file visit the above url
\item bimbam: Mean genotype file format of bimbam which is a single file, 
without information on individuals. From 
\url{http://www.haplotype.org/bimbam.html}: The first column of the mean 
genotype files is the SNP ID, the second and third columns are allele types 
with minor allele first. The rest columns are the mean genotypes of different 
individuals – numbers between 0 and 2 that represents the (posterior) mean 
genotype, or dosage of the minor allele.
}
}

